# apps/cms/openapi.yaml
openapi: 3.1.0
info:
  title: LumiBase CMS API
  version: 0.1.0
  description: |
    Edge-native Headless CMS API. See `docs/api/hono-api-spec.md` for the full
    surface; this file is the machine-readable contract that ships with the
    repo. Endpoints land incrementally — Phase 0 covers auth/health only.
servers:
  - url: https://api.lumibase.dev/api/v1
    description: Production
  - url: http://127.0.0.1:1989/api/v1
    description: Local wrangler dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SiteHeader:
      name: X-Lumi-Site
      in: header
      required: true
      schema: { type: string }
      description: Active tenant id. Required for every authenticated endpoint.
  schemas:
...
    Relation:
      type: object
      required: [id, siteId, manyCollection, manyField, oneCollection]
      properties:
        id: { type: string }
        siteId: { type: string }
        manyCollection: { type: string }
        manyField: { type: string }
        oneCollection: { type: string }
        oneField: { type: string, nullable: true }
        junctionCollection: { type: string, nullable: true }
        type: { type: string }
        aliasField: { type: string, nullable: true }
        relatedDisplayTemplate: { type: string, nullable: true }
        junctionManyField: { type: string, nullable: true }
        junctionOneField: { type: string, nullable: true }
        sortField: { type: string, nullable: true }
        onDelete: { type: string, enum: [restrict, cascade, set null, no action] }
        meta: { type: object, additionalProperties: true }
...
paths:
...
  /relations:
    get:
      summary: List relations in the active site
      parameters: [{ $ref: '#/components/parameters/SiteHeader' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Relation' }
    post:
      summary: Create a relation
      parameters: [{ $ref: '#/components/parameters/SiteHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Relation' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Relation' }
  /relations/{id}:
    parameters:
      - $ref: '#/components/parameters/SiteHeader'
      - { name: id, in: path, required: true, schema: { type: string } }
    delete:
      summary: Delete a relation
      responses:
        '204': { description: Deleted }
...
  /items/{collection}:
    parameters:
      - $ref: '#/components/parameters/SiteHeader'
      - { name: collection, in: path, required: true, schema: { type: string } }
    get:
      summary: List items in a collection
      parameters:
        - { name: fields, in: query, schema: { type: string } }
        - { name: filter, in: query, schema: { type: string }, description: 'JSON filter DSL' }
        - { name: sort, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0 } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        '200': { description: OK }
    post:
      summary: Create an item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data]
              properties:
                data: { type: object }
                status: { type: string }
                sort: { type: integer }
      responses:
        '201': { description: Created }
  /items/{collection}/bulk:
    parameters:
      - $ref: '#/components/parameters/SiteHeader'
      - { name: collection, in: path, required: true, schema: { type: string } }
    post:
      summary: Bulk operation (create/update/delete) on items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [op, items]
              properties:
                op: { type: string, enum: [create, update, delete] }
                items:
                  type: array
                  items: { type: object }
      responses:
        '200': { description: OK }
  /items/{collection}/{id}:
    parameters:
      - $ref: '#/components/parameters/SiteHeader'
      - { name: collection, in: path, required: true, schema: { type: string } }
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      summary: Get an item
      responses: { '200': { description: OK } }
    patch:
      summary: Patch an item (merge data)
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses: { '200': { description: OK } }
    put:
      summary: Replace item data
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses: { '200': { description: OK } }
    delete:
      summary: Soft-delete an item
      responses: { '204': { description: Deleted } }
  /items/{collection}/{id}/revisions:
    parameters:
      - $ref: '#/components/parameters/SiteHeader'
      - { name: collection, in: path, required: true, schema: { type: string } }
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      summary: List revisions for an item
      responses: { '200': { description: OK } }
  /items/{collection}/{id}/revert/{revisionId}:
    parameters:
      - $ref: '#/components/parameters/SiteHeader'
      - { name: collection, in: path, required: true, schema: { type: string } }
      - { name: id, in: path, required: true, schema: { type: string } }
      - { name: revisionId, in: path, required: true, schema: { type: string } }
    post:
      summary: Revert an item to a previous revision snapshot
      responses: { '200': { description: OK } }
...